Added ValidationBookSolver to skip permutaions which contain invalid connections. Updated to Kotlin 1.6.20 Replaced com.nhaarman.mockitonkotlin2 with org.mockito.kotlin

+++ abr-core/src/main/kotlin/com/d20charactersheet/adventurebookresolver/core/domain/BookSolverListener.kt
@@ -8,10 +8,10 @@ interface BookSolverListener {
 
     fun endCalculation(endTime: LocalDateTime)
 
-    fun calculateCombinations(numberOfCombinations: Int)
+    fun calculateCombinations(numberOfCombinations: Long)
 
     fun calculatePath(startEntry: BookEntry, wayPoint: BookEntry, numberOfEntries: Int?)
 
-    fun maxCombinations(maxCombinations: Int)
+    fun maxCombinations(maxCombinations: Long)
 
 }
